+++ ndh/forms.py
@@ -1,16 +1,16 @@
 """General Form-related helpers for NDH."""
 from django import forms
 
-from .widgets import DatalistInput
+from .widgets import DATE_FORMAT, TYPE_DATE, TYPE_TIME, DatalistInput
 
 
 class AccessibleDateTimeField(forms.SplitDateTimeField):
     """Set date & time HTML types."""
 
     widget = forms.SplitDateTimeWidget(
-        date_format="%Y-%m-%d",
-        date_attrs={"type": "date"},
-        time_attrs={"type": "time"},
+        date_format=DATE_FORMAT,
+        date_attrs=TYPE_DATE,
+        time_attrs=TYPE_TIME,
     )  # type: ignore